Moved the utility functions to their own file so they can be used independently of main.py. Imports are what is used in each file (python handles multiple imports gracefully)
